requirements.txt
fastapi
uvicorn
tensorflow==2.12.0  # If you're using Python 3.10 or lower
pillow
numpy
